Do no lock the audio device in Resamplers or Decoders
Only Stream is responsible for playback. Resamplers and Decoders should
not even assume that an audio device is even available, let alone trying
to lock it. Furthermore, the SDL audio callback is calling Resampler and
Decoder functions directly, and locking the audio device from within the
audio callback could be problematic.