Force consistent delay for stream playback (#26)
The idea here is to force the delay for stream playback to be equal to
the length of the audio buffer. This is achieved by applying an offset
to the first frame of playback based on the amount of time elapsed
between the start of playback and the subsequent audio frame callback.

This provides a more consistent delay and smooths out the unevenness in
playback that comes about from always aligning the start of each stream
to the start of the next audio buffer.