Fix clicks/pops when seeking with DecoderVorbis
ov_time_seek_lap() and ov_raw_seek_lap() are great when doing live-seeks
on currently playing streams. However, they're not so great for non-live
seeks. A stopped stream, or a stand-alone DecoderVorbis will produce
clicks and pops after seeking.

Change to using ov_time_seek() and ov_raw_seek() to solve the issue. The
drawback now of course is that live-seeking can be more clicky. However,
other decoders (like libmpg123 and opusfile) don't even attempt to do
transition crosslapping while seeking, and they're always clicky. So
it's questionable if this is something that should be even offered at
the API level of a decoder. It seems to me softening seek transitions
is something to be done at the application level, not by the decoder.