Fix lib and pkgconfig install location
It's not specified whether CMAKE_INSTALL_LIBDIR a relative path or not.
CMake is supposed to do some magic with it depending on how you use it,
but clearly that's not always working. The lib suffix might be wrong
(like "lib32" instead of "lib", or "lib" instead of "lib64") or on some
systems we get something silly like "/usr/local/usr/local/lib" when
combined with CMAKE_INSTALL_PREFIX. See:

https://bugzilla.redhat.com/show_bug.cgi?id=795542

The GNUInstallDirs module defines CMAKE_INSTALL_FULL_LIBDIR, which is
guaranteed to be absolute. So just use that instead.