Support using SDL_audiolib with `-fno-exceptions`
When `SDL_audiolib` is included as a subproject and the parent project
disables C++ exceptions, it previously failed to compiled due to a
single `throw` call in `SdlMutex.h`.

Extracts `SdlMutex.cpp.o` into a separate object library that always has
`-fexceptions` to allow building the rest of `SDL_audiolib` without
exceptions.